+++ src/main/java/com/tagnumelite/projecteintegration/addons/WootAddon.java
@@ -23,6 +23,8 @@
 package com.tagnumelite.projecteintegration.addons;
 
 import com.tagnumelite.projecteintegration.api.recipe.APEIRecipeMapper;
+import com.tagnumelite.projecteintegration.api.recipe.nss.NSSInput;
+import com.tagnumelite.projecteintegration.api.recipe.nss.NSSOutput;
 import ipsis.woot.crafting.AnvilRecipe;
 import ipsis.woot.crafting.FluidConvertorRecipe;
 import ipsis.woot.crafting.InfuserRecipe;